+++ diive/core/times/resampling.py
@@ -14,7 +14,7 @@ def resample_series_to_30MIN(series: Series,
                              to_freqstr: str = '30T',
                              agg: str = 'mean',
                              mincounts_perc: float = .9,
-                             output_timestamp_shows: Literal['middle', 'end'] = 'end'):
+                             output_timestamp_shows: Literal['middle', 'end'] = 'end') -> Series:
     """Downsample data to 30-minute time resolution
 
     Input data must have timestamp showing the END of the time period.
@@ -43,6 +43,9 @@ def resample_series_to_30MIN(series: Series,
 
     """
 
+    current_freq = to_offset(series.index.freqstr)
+    requested_freq = to_offset(to_freqstr)
+
     # Resampling only 30MIN time resolution
     if not any(chr in to_freqstr for chr in ['30T']):
         raise NotImplementedError("Error during resampling: Only resampling to 30 minutes (30T) allowed.")
@@ -51,13 +54,8 @@ def resample_series_to_30MIN(series: Series,
     if not series.index.freq:
         raise NotImplementedError("Error during resampling: Irregular timestamps are not supported.")
 
-    # Requested frequency must be different from data freq
-    # 'to_offset' allows comparison of frequencies (larger, smaller, equal)
-    if to_offset(series.index.freqstr) == to_offset(to_freqstr):
-        return series
-
     # Requested frequency must be larger than data freq
-    if to_offset(series.index.freqstr) > to_offset(to_freqstr):
+    if current_freq > requested_freq:
         raise NotImplementedError(
             f"Error during resampling: "
             f"Upsampling not allowed. "
@@ -66,10 +64,11 @@ def resample_series_to_30MIN(series: Series,
 
     _series = series.copy()
 
-    print(f"Resampling data from {series.index.freqstr} to {to_freqstr} frequency ...")
-
     # Make middle timestamp, for correct resampling
     _series = convert_series_timestamp_to_middle(data=_series)
+
+    print(f"Resampling data from {current_freq.freqstr} "
+          f"to {requested_freq.freqstr} frequency ...")
 
     # Check maximum number of counts per aggregation interval
     # Needed to calculate the required minimum number of counts from
@@ -79,25 +78,40 @@ def resample_series_to_30MIN(series: Series,
     # maxcounts = maxcounts.resample(to_freqstr, label=label, closed=closed).count().max()
     mincounts = int(maxcounts * mincounts_perc)
 
+    # Require minimum 1 value
+    # Relevant e.g. for 10MIN data (3 values per half-hour)
+    mincounts = 1 if mincounts < 3 else mincounts
+
     # Aggregation
-    resampled_df = _series.resample(to_freqstr, label='right')  # default: closed='left'
+    resampled_ser = _series.resample(to_freqstr, label='right')  # default: closed='left'
     # resampled_df = _series.resample(to_freqstr, label=label, closed=closed)
-    agg_counts_df = resampled_df.count()  # Count aggregated values, always needed
-    agg_df = resampled_df.agg(agg)
+    agg_counts_ser = resampled_ser.count()  # Count aggregated values, always needed
+    agg_ser = resampled_ser.agg(agg)
 
     # Timestamp index shows end after resampling b/c label='right'
-    agg_counts_df.index.name = 'TIMESTAMP_END'
-    agg_df.index.name = 'TIMESTAMP_END'
+    agg_counts_ser.index.name = 'TIMESTAMP_END'
+    agg_ser.index.name = 'TIMESTAMP_END'
 
     # Keep aggregates with enough values
-    filter_min = agg_counts_df >= mincounts
-    agg_df = agg_df[filter_min]
+    filter_min = agg_counts_ser >= mincounts
+    agg_ser = agg_ser[filter_min]
+
+    # Re-assign 30MIN resolution as freq
+    agg_ser = agg_ser.asfreq(to_freqstr)
+
+    # elif current_freq == requested_freq:
+    #     # Requested frequency must be different from data freq
+    #     # 'to_offset' allows comparison of frequencies (larger, smaller, equal)
+    #     print(f"No resampling needed, data freq {current_freq} "
+    #           f"is already in requested freq {requested_freq} ...")
+    #     agg_ser = _series.copy()
 
     # Sanitize resampled timestamp index
-    if output_timestamp_shows == 'middle':
-        agg_df = TimestampSanitizer(data=agg_df).get()
-    elif output_timestamp_shows == 'end':
-        agg_df = TimestampSanitizer(data=agg_df, output_middle_timestamp=False).get()
+    if not agg_ser.empty:
+        if output_timestamp_shows == 'middle':
+            agg_ser = TimestampSanitizer(data=agg_ser).get()
+        elif output_timestamp_shows == 'end':
+            agg_ser = TimestampSanitizer(data=agg_ser, output_middle_timestamp=False).get()
 
     # agg_df = sanitize_timestamp_index(data=agg_df, freq='30T')
 
@@ -115,4 +129,4 @@ def resample_series_to_30MIN(series: Series,
     #                                                  out_timestamp_convention='Middle of Record')
     # agg_df.index = pd.to_datetime(agg_df.index)
 
-    return agg_df
+    return agg_ser